Refactor game ideation and README to align FPS Tank Commander objectives, implement Asteroid game page move, added crosshair component and sound, and update game links in index.
