+++ app/pages/asteroid/components/UI/Crosshair.module.css
@@ -1,0 +1,37 @@
+.crosshairWrapper {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  width: 32px;
+  height: 32px;
+  transform: translate(-50%, -50%);
+  z-index: 3;
+  pointer-events: none;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.crosshairVertical {
+  width: 2px;
+  height: 20px;
+  background: red;
+  position: absolute;
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%);
+  border-radius: 1px;
+  opacity: 0.8;
+}
+
+.crosshairHorizontal {
+  width: 20px;
+  height: 2px;
+  background: red;
+  position: absolute;
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%);
+  border-radius: 1px;
+  opacity: 0.8;
+}
